+++ fabric/src/main/java/house/greenhouse/bovinesandbuttercups/content/recipe/ingredient/BovinesIngredients.java
@@ -1,0 +1,9 @@
+package house.greenhouse.bovinesandbuttercups.content.recipe.ingredient;
+
+import net.fabricmc.fabric.api.recipe.v1.ingredient.CustomIngredientSerializer;
+
+public class BovinesIngredients {
+    public static void init() {
+        CustomIngredientSerializer.register(RemainderIngredientImpl.SERIALIZER);
+    }
+}